requirements.txt
python-dotenv
urllib3
beautifulsoup4
Pillow
opencv-python
Wand
ffmpeg-python
deep-translator
google-api-python-client
google-cloud-storage
Google-Images-Search
saucenao_api
ShazamAPI
captcha
price-parser
spamwatch
openai
Pylint